The password generator to generates the order of letters and symbols randomly now. I added some output to the build.sh shell script and I will add kwargs in the future

+++ build.sh
@@ -1,4 +1,14 @@
-rm -rf dist/*
+#!/bin/bash
+
+red='\e[1;31m%s\e[0m\n'
+green='\e[1;32m%s\e[0m\n'
+yellow='\e[1;33m%s\e[0m\n'
+blue='\e[1;34m%s\e[0m\n'
+magenta='\e[1;35m%s\e[0m\n'
+cyan='\e[1;36m%s\e[0m\n'
+
+[ -d "dist/" ] && rm -rf dist/* || printf "$green" "dist directory does not exist!"
+[ -f "MANIFEST" ] && rm -f MANIFEST || printf "$green" "MANIFEST does not exist"
 python3 setup.py sdist
-/home/$(whoami)/.local/bin/twine upload dist/*
+twine upload dist/*
